Move IR/pipeline related command/messages
